+++ resources/release-notes.md
@@ -1,5 +1,6 @@
 # Software
 
 - Add comparison view `mag.graphics.view.Comparison` for comparing science, I-ALiRT and SID5 data
-- Improve management of compression in `mag.process.SignedInteger`
 - Add `RangeVariable` as option in `mag.process.Calibration`
+- Improve management of compression in `mag.process.SignedInteger`
+- Fix issues with processing of SID5 data (signedness, calibration)